chore: set initial loan fee to one basis point (#2835)
one basis point fee corresponds with a daily interest charge that
would accrue to 3.5% APR.  This level means that the minimum loan size
is 10000 RUN, because any less would get an initial fee of 0 RUN,
which is currently disallowed. The previous setting (50 basis points)
alllowed for a minimum loan of 200.