fix(agoric-cli): increase integration-test timeout
Now that spawner uses createVat() (#1343), deploys to the solo machine take
longer. The integration test was allowing 60s for `agoric deploy
./contract/deploy.js ./api/deploy.js` to work, and that used to take 47s on
my local machine, but now it takes 76s, and would exceed the timeout.

I increased the timeout to 120s to fix this.

When we switch vats to use XS, this should be dramatically faster (about
10x), so we'll have have plenty of margin.