fix: remove deprecated ibid support (#2898)
Part 2 of the removal of ibids. #2896 changed our code to no longer send ibids, but still understood received ibids to not break and near cross-version communications. More than a week after #2896 was merged, this PR removes the code for understanding them (except to give a more informative error if it sees one anyway). After this PR we should still be compatible with senders running code after #2896. But not earlier.