fix(swingset): add GC stubs: syscall/dispatch retireImports/Exports
This add `syscall.retireImports` and `syscall.retireExports` to the
pre-existing `syscall.dropImports`. All three are no-ops, but vats should be
able to call them without ill effects.

It also adds `dispatch.retireExports` and `dispatch.retireImports` to the
pre-existing `dispatch.dropExports` stub. These are tolerated but ignored by
both liveslots and the comms vat

Together, this should allow kernel and liveslots GC development to proceed in
parallel.

refs #2724