fix(cosmic-swingset): slog begin/end-block and input events
Changes cosmic-swingset to add the following events to the slogfile:

* `type: 'cosmic-swingset-deliver-inbound'`: an inbound message was given to
  the mailbox device
* `type: 'cosmic-swingset-bridge-inbound'`: a message was given to the bridge
  device
* `type: 'cosmic-swingset-begin-block'`: the cosmos-side SwingSet module was
  given a BEGIN_BLOCK event
* `type: 'cosmic-swingset-end-block-start'`: the cosmos-side SwingSet module
  was given an END_BLOCK event, and SwingSet's `controller.run()` is about to
  be called
* `type: 'cosmic-swingset-end-block-finish'`: the cosmos-side SwingSet module
  was given an END_BLOCK event, and SwingSet's `controller.run()` has just
  finished running

The begin/end block events include the block's height and time, both of which
are part of the consensus state. The deliver/bridge events include some
information about the messages being delivered.

This enables slogfile post-processing tools to correlate cranks with the
blocks whose transactions provoked them.