fix(SwingSet): let vatManager creator override syscall comparison
In normal operation, any syscall divergence during replay is a fatal
"anachrophobia error". But when we're replaying a vat for local debugging
purposes, we might want to be more tolerant, e.g. Errors being serialized
differently when the worker is in a debug mode.

This add a `managerOptions.compareSyscalls` which can override the way
transcript replay works.

We export `requireIdentical()` so callers can use it to build their custom
`compareSyscalls()` function.