feat(xsnap): add gcAndFinalize, tests
This changes the xsnap.c run loop to give finalizers a chance to run just
after the promise queue drains.

With this change, userspace can do a combination of `gc()` and `setImmediate`
that lets it provoke a full GC sweep, and wait until finalizers have run.
SwingSet will use this during a crank, after userspace has become idle, and
before end-of-crank GC processing takes place.

This combination is implemented in a function named `gcAndFinalize()`. We
copy this function from its normal home in SwingSet so the xsnap.c behavior
it depends upon can be tested locally.

refs #2660