fix(swingset): track pendingPromises
We don't do GC tracking of Promises, since we don't really have a good theory
of how it should work. However, most Promises get resolved, and most are not
mentioned after resolution. So we save memory by dropping the Promise after
it gets resolved, retiring its identity in the process.

To make this safe, we must ensure those unresolved Promises stay in memory
until they get resolved, even though userspace may or may not retain a strong
reference to them. So for every Promise that liveslots exports, and for every
Promise liveslots creates during import, we store a strong reference in
`pendingPromises` until it is resolved and retired (i.e. we add the Promise
to the Set). This strongref is dropped in `retirePromiseID`.

This probably isn't strictly necessary for imported promises, because
`importedPromisesByPromiseID` retains a strong reference to the
resolve/reject functions, and experimental evidence shows that either one
will keep the Promise itself in memory. But it's not clear that this is part
of the JS spec. So to be safe, we use `pendingPromises` to hold onto the
Promise directly.

For exported promises, liveslots calls `p.then` to watch for resolution (so
it can inform the kernel if/when/how it is resolved). From my experiments,
this is *not* sufficient to keep the Promise alive, hence `pendingPromises`
is important.

The potential failure mode (exercised by the accompanying unit test) is if
userspace created a unresolveable Promise (i.e. it dropped the `resolve` and
`reject` functions) and exported it to the kernel, then the kernel sends the
promise vref back into the vat. If we didn't keep a strongref, the slotToVal
WeakRef would die, and liveslots would be unable to map the vref back to a
Promise object (and would refuse to make an imported Promise for it, because
the allocation flag on the vref shows that the vat allocated it originally).