chore(vats): downgrade "cannot create fee collector" error message
@michaelfig suggested that I downgrade the "Cannot create fee collector"
error message from `console.error` to `console.log`. It happens just after
the Bank vat throws a "Bank doesn't implement fee collectors" Error when its
`getFeeCollectorDepositFacet()` method is invoked. The Error is written to
stdio as soon as it's serialized (xsnap/SES does that with everything), and
we aren't changing that. This patch only changes a `console.error` that sat
in the `.catch()` clause on that message. But at least it removes one of the
15 lines from the log.