feat(swing-store-simple)!: remove .jsonlines hack from simple swing store
BREAKING CHANGE: most uses of simple swing store use it to get a purely
in-memory store for testing purposes, but a few places used the older .jsonlines
thing for simple persistence.  Any existing code that did that requires revision
to use swing-store-lmdb or another persistence mechanism, though all the known
cases that did that have been so revised as part of this update.