fix(swingset): use provideVatSlogger inside the slogger
The old code implemented the various vat-specific slogger methods by using
`vatSlogs.get(vatID)`, which assumes that someone has already created the
vatSlogger. The new+correct approach is to call `provideVatSlogger` each
time, but the slogger internals had not been updated. This fixes the problem,
avoiding a crash when slogging is enabled and a vat operation is performed in
a particular order.