fix(swingset): tolerate policy='none' in queueToVatExport
I added this mode (which doesn't even bother allocating a result promise) in
94c1688583a3a7510d329d3166431eb97cdabb7f, but I didn't change
queueToVatExport to tolerate the lack of a result promise. This fixes the
issue. Unit tests that want to provoke an action but don't care about
observing the result (and want to minimize the number of additional cranks
being executed) should use `'none'`.