chore(swingset): enable unregister() in liveslots
To prepare for removing things from the FinalizationRegistry (specifically
exports that have been retired by the kernel, via `dispatch.retireExports()`,
but which still have a live Remotable and are thus still tracked in the
slotToVal table and `droppedRegistry`), register these objects with the
object itself as the "unregister handle". This will allow
`droppedRegister.unregister()` to work in a future change.