feat(swingset): comms: add/manipulate isReachable flag
Each c-list object entry now has an `isReachable` flag. On import-facing
c-lists, this indicates that the downstream (kernel or remote) can reach the
object (and is cleared when the downstream side drops their strong
reference). On the one export-facing c-list, we use the flag to remember
whether we've dropped our import or not.

c-list translators now ensure the flag is set after an allocation-worthy
reference (imports during inbound translation, exports during outbound
translation) is converted. C-list entry deletion functions ensure the flag is
clear. In either case, if the state of the flag actually changed (on an
import-side c-list), the `reachable` refcount is modified. If that refcount
touches zero, the object is marked as "maybe free" for later investigation.

`addEgress` also sets the isReachable flag, to reflect the initial conditions
of a manually-added c-list entry.