fix(swingset): test-comms.js: fix retireImports test
This unit test was exercising `dispatch.retireImports`, just to ensure that
it didn't cause a crash. But the invocation was not semantically valid: it
was retiring an object without first dropping it. That didn't matter when the
implementation was a no-op, but now that it actually behaves correctly, this
invocation is invalid, and causes an assertion failure.

The new test exercises this correctly, at least in one direction.