fix(swingset): comms: single-argument delete-mapping functions
The comms vat c-list management code has a pair of
functions (`deleteRemoteMapping` and `deleteKernelMapping`) to remove c-list
entries. These functions take both a local reference (lref) and an
externally-facing reference (kfref for the kernel-facing side, rref for the
remote-facing side).

This change removes the externally-facing reference from the API, so now you
call it with only the lref. Both functions look up the necessary reference
internally.

This removes an easy-to-make error from `deleteRemoteMapping()`, because
rrefs come in both inbound- and outbound- flavors, and calling
`deleteRemoteMapping()` with the wrong flavor would leave half the c-list
entry undeleted.